Rename module streamed_str to streaming
