+++ src/magentic/__init__.py
@@ -1,7 +1,7 @@
 from magentic.function_call import FunctionCall
 from magentic.prompt_chain import prompt_chain
 from magentic.prompt_function import prompt
-from magentic.streamed_str import AsyncStreamedStr, StreamedStr
+from magentic.streaming import AsyncStreamedStr, StreamedStr
 
 __all__ = [
     "FunctionCall",
