+++ src/magentic/streaming.py
