+++ tests/test_prompt_function.py
@@ -8,7 +8,7 @@ from pydantic import BaseModel
 from magentic.chat_model.openai_chat_model import StructuredOutputError
 from magentic.function_call import FunctionCall
 from magentic.prompt_function import AsyncPromptFunction, PromptFunction, prompt
-from magentic.streamed_str import AsyncStreamedStr, StreamedStr
+from magentic.streaming import AsyncStreamedStr, StreamedStr
 
 
 @pytest.mark.openai
