+++ tests/test_streaming.py
